+++ code/Universe/Interfaces/IGalaxy.cs
@@ -3,33 +3,8 @@
 namespace Universe.Interfaces;
 
 /// <summary></summary>
-public interface IGalaxy<T> where T : ICosmicEntity
+public interface IGalaxy<T> : IGalaxyBasic<T> where T : ICosmicEntity
 {
-    /// <summary>
-    /// Create a new model in the database
-    /// </summary>
-    Task<(Gravity g, string t)> Create(T model);
-
-    /// <summary>
-    /// Bulk create new models in the database
-    /// </summary>
-    Task<Gravity> Create(IList<T> models);
-
-    /// <summary>
-    /// Modify a model in the database
-    /// </summary>
-    Task<(Gravity g, T T)> Modify(T model);
-
-    /// <summary>
-    /// Bulk modify models in the database
-    /// </summary>
-    Task<Gravity> Modify(IList<T> models);
-
-    /// <summary>
-    /// Remove one model from the database
-    /// </summary>
-    Task<Gravity> Remove(string id, string partitionKey);
-
     /// <summary>
     /// Get one model from the database
     /// </summary>
@@ -49,4 +24,9 @@ public interface IGalaxy<T> where T : ICosmicEntity
     /// Get a paginated list from the database
     /// </summary>
     Task<(Gravity g, IList<T> T)> Paged(Q.Page page, IList<Cluster> clusters, ColumnOptions? columnOptions = null, IList<Sorting.Option> sorting = null, IList<string> group = null);
+
+    /// <summary>
+    /// Execute a stored procedure
+    /// </summary>
+    Task<(Gravity g, T T)> ExecSProc(string procedureName, string partitionKey, params object[] parameters);
 }
