+++ code/Universe/global-usings.cs
@@ -5,6 +5,6 @@ global using System.Text;
 global using System.Text.Json.Serialization;
 global using System.Threading.Tasks;
 global using Microsoft.Azure.Cosmos;
+global using Universe.Builder.Options;
 global using Universe.Exception;
 global using Universe.Interfaces;
-global using Universe.Options.Query;
